Rename plugin to 'notebooks'
