--- plugins/jupyter-integration/gradle.properties
@@ -1,27 +1,0 @@
-#
-# Copyright 2019-2023 The Polypheny Project
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-pluginVersion = 0.0.1
-
-pluginId = jupyter-integration
-pluginClass = org.polypheny.db.jupyter.JupyterPlugin
-pluginProvider = The Polypheny Project
-pluginDependencies = sql-language, mql-language, cypher-language
-pluginUrlPath =
-pluginCategories =
-pluginPolyDependencies =
-pluginIsSystemComponent = false
-pluginIsUiVisible = true
