--- plugins/jupyter-integration/src/main/java/org/polypheny/db/jupyter/model/request/ContentsCreateRequest.java
@@ -1,30 +1,0 @@
-/*
- * Copyright 2019-2023 The Polypheny Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.polypheny.db.jupyter.model.request;
-
-import com.google.gson.annotations.SerializedName;
-import lombok.Getter;
-
-@Getter
-public class ContentsCreateRequest {
-
-    @SerializedName("copy_from")
-    private String copyFrom;
-    private String ext;
-    private String type;
-
-}
