+++ plugins/notebooks/src/main/java/org/polypheny/db/notebooks/model/language/JupyterLanguageFactory.java
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2019-2023 The Polypheny Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.polypheny.db.notebooks.model.language;
+
+public class JupyterLanguageFactory {
+
+    /**
+     * Get the JupyterKernelLanguage implementation for the given kernel
+     *
+     * @param kernelName the name that uniquely identifies a kernel
+     * @return a JupyterKernelLanguage for the specified language or null if the language is not supported
+     */
+    public static JupyterKernelLanguage getKernelLanguage( String kernelName ) {
+        if ( kernelName == null ) {
+            return null;
+        }
+        if ( kernelName.equalsIgnoreCase( "python3" ) ) {
+            return new IPythonKernelLanguage();
+        }
+        // add supported kernels here
+
+        return null;
+    }
+
+}
