+++ plugins/notebooks/src/main/java/org/polypheny/db/notebooks/model/request/ContentsSaveRequest.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2019-2023 The Polypheny Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.polypheny.db.notebooks.model.request;
+
+import lombok.Getter;
+
+@Getter
+public class ContentsSaveRequest<T> {
+
+    private T content;
+    private String format;
+    private String name;
+    private String path;
+    private String type;
+
+}
