+++ settings.gradle
@@ -48,7 +48,7 @@ include 'plugins:excel-adapter'
 
 // other plugins
 include 'plugins:explore-by-example'
-include 'plugins:jupyter-integration'
+include 'plugins:notebooks'
 
 // disabled adapter plugins
 include 'plugins:pig-adapter'
